requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pyjwt==2.11.0
